[REFACTOR] : useFetchStreamData 수정
